MVCF: Move trackers to own folder
