+++ Source/MVCF/Features/Feature_ApparelVerbs.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
-using MVCF.PatchSets;
+using MVCF.PatchSets.Trackers;
 
 namespace MVCF.Features;
