+++ Source/MVCF/PatchSets/Trackers/PatchSet_Equipment.cs
@@ -6,7 +6,7 @@ using MVCF.ModCompat;
 using MVCF.Utilities;
 using Verse;
 
-namespace MVCF.PatchSets;
+namespace MVCF.PatchSets.Trackers;
 
 public class PatchSet_Equipment : PatchSet
 {
