+++ Source/MVCF/PatchSets/Trackers/PatchSet_Inventory.cs
@@ -5,7 +5,7 @@ using MVCF.Comps;
 using MVCF.Utilities;
 using Verse;
 
-namespace MVCF.PatchSets;
+namespace MVCF.PatchSets.Trackers;
 
 public class PatchSet_Inventory : PatchSet
 {
